+++ lib/rubocop.rb
@@ -50,6 +50,7 @@ require 'rubocop/cop/alias'
 require 'rubocop/cop/rescue_modifier'
 require 'rubocop/cop/percent_literals'
 require 'rubocop/cop/brace_after_percent'
+require 'rubocop/cop/ensure_return'
 
 require 'rubocop/report/report'
 require 'rubocop/report/plain_text'