+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * New cop `PercentLiterals` tracks uses of `%q`, `%Q`, `%s` and `%x`
 * New cop `BraceAfterPercent` tracks uses of % literals with delimiters other than ()
 * Support for disabling cops locally in a file with rubocop:disable comments.
+* New cop `EnsureReturn` tracks usages of `return` in `ensure` blocks.
 
 ### Bugs fixed